+++ e-arbiter-tournament/src/main/kotlin/pl/cyganki/tournament/model/dto/TaskPreview.kt
@@ -5,6 +5,6 @@ import pl.cyganki.utils.model.TaskUserDetails
 data class TaskPreview(
         val name: String,
         val description: String,
-        val maxPoints: Long,
+        val maxPoints: Int,
         val taskUserDetails: TaskUserDetails?
 )
